+++ examples/example_diffusion.R
@@ -1,9 +1,9 @@
 # example for diffusion() function
 fitbass <- diffusion(tschicken[, 2], type = "bass")
 fitgomp <- diffusion(tschicken[, 2], type = "gompertz")
-fitsgomp <- diffusion(tschicken[, 2], type = "sgompertz")
+fitgsg <- diffusion(tschicken[, 2], type = "gsgompertz")
 
 # Produce some plots
 plot(fitbass)
 plot(fitgomp)
-plot(fitsgomp)
+plot(fitgsg)