+++ examples/example_nortonbass.R
@@ -1,21 +1,21 @@
 ### Norton Bass
 
 \dontrun{
-  data(ibm)
+  data(tsibm)
   
-  fitNB1 <- Nortonbass(ibm, startval.met = "2ST", estim.met = "OLS",
+  fitNB1 <- Nortonbass(tsibm, startval.met = "2ST", estim.met = "OLS",
                        startval = NULL, flexpq = F, gstart = NULL)
-  fitNB2 <- Nortonbass(ibm, startval.met = "2ST", estim.met = "OLS",
-                       startval = NULL, flexpq = T, gstart = NULL)
+  fitNB2 <- Nortonbass(tsibm, startval.met = "2ST", estim.met = "SUR",
+                       startval = NULL, flexpq = F, gstart = NULL)
   # using BOBYQA algorithm
-  fitNB3 <- Nortonbass(ibm, startval.met = "2ST", estim.met = "BOBYQA",
+  fitNB3 <- Nortonbass(tsibm, startval.met = "2ST", estim.met = "BOBYQA",
                        startval = NULL, flexpq = F, gstart = NULL)
   
   # Create some plots
-  plot(ibm[, 1],type = "l", ylim=c(0,35000))
-  lines(ibm[, 2],col ="blue")
-  lines(ibm[, 3],col ="green")
-  lines(ibm[, 4],col ="pink")
+  plot(tsibm[, 1],type = "l", ylim=c(0,35000))
+  lines(tsibm[, 2],col ="blue")
+  lines(tsibm[, 3],col ="green")
+  lines(tsibm[, 4],col ="pink")
   
   lines(fitNB1$fit$fitted[[1]], col = "black", lty = 2)
   lines(fitNB1$fit$fitted[[2]], col = "blue", lty = 2)
@@ -33,15 +33,15 @@
   lines(fitNB3$fit$fitted[[4]], col = "pink", lty = 4)
   
   # read out RMSE
-  fitNB1$fitted[[1]]$rmse
-  fitNB1$fitted[[2]]$rmse
-  fitNB1$fitted[[3]]$rmse
-  fitNB1$fitted[[4]]$rmse
+  fitNB1$fit$RMSE[[1]]
+  fitNB1$fit$RMSE[[2]]
+  fitNB1$fit$RMSE[[3]]
+  fitNB1$fit$RMSE[[4]]
   
-  fitNB2$fitted[[1]]$rmse
-  fitNB2$fitted[[2]]$rmse
-  fitNB2$fitted[[3]]$rmse
-  fitNB2$fitted[[4]]$rmse
+  fitNB2$fit$RMSE[[1]]
+  fitNB2$fit$RMSE[[2]]
+  fitNB2$fit$RMSE[[3]]
+  fitNB2$fit$RMSE[[4]]
   
   fitNB3$fit$RMSE[[1]]
   fitNB3$fit$RMSE[[2]]